Refactor code for better readability and debugging
The changes in this commit aimed to improve code readability and debugging. In `SmsTest.php`, the format of the `testSmsIsRegister` function was updated to adhere to the PSR-2 coding standard.

In `hyperf.php`, the `Debugger::runOnTTY();` call was changed to instantiate an object and provide a console output for better debugging. This output will be shown in yellow color for highlight.