Refactor error reporting in WebSocketDebugger
In WebSocketDebugger script, the way exceptions and errors were handled has been refactored. Instead of using the separate methods 'exception' and 'error', that functionality has moved to the 'send' method for more direct error reporting. A new import 'WeakMap' has been added although its use is not reflected in the changes, it could be required in farther updates.