"Remove unnecessary newline in WebSocketDebugger's payload data"
In this version of the file, we have removed the PHP_EOL which was adding a newline character at the end of payload data in WebSocketDebugger.php. This tweak was necessary as this newline character was causing unnecessary line breaks and presenting the payload data in a split and confusing manner in the client-side.