Remove TcpDebugger and add SSL support to WebSocketDebugger
TcpDebugger file was removed as it was found unnecessary and was not being used further in the project. Implemented SSL Support in WebSocketDebugger which allows the application to encrypt the traffic between the server and client for enhanced security. Updated the SslConfig file to not validate the given certificate and certificate key, instead, it only fetches it. This reduces burden on the SslConfig and shifts the responsibility of validation to where it's needed, providing a better project organization and avoiding unnecessary operations. These changes are made for maintaining cleaner code and enhancing security measures.