Add detailed Swow-Debugger usage instructions to README
This commit adds an extensive list of Swow-Debugger tools and how to use them to the README file. The extensive list provides details on how to use commands in Swow-Debugger for activities such as viewing current coroutines statuses, debugging step by step, setting breakpoints, viewing and modifying the values of variables, killing coroutines, scanning for zombie coroutines, and many more. This addition is intended to make it easier for users to understand and utilize Swow-Debugger effectively.