Refactor pool handling and add viewing online configuration command
This commit refactors the way the pool command is handled in WebSocketDebugger and adds a new feature - viewing online configuration information. Instead of performing manual checks on the type of pool ('redis' or 'mysql'), a more expressive `match` statement is used. This provides a more robust solution. In addition, a 'config' command has been introduced, which allows users to view online configuration information. This addition improves the debugging capabilities of the software. The README file has been updated correspondingly with the usage of the new 'config' command. Note: sensitive information fetching via 'config' command is not recommended.