Add 'route' case to WebSocketDebugger
In the switch statement of the WebSocketDebugger class, an additional case for 'route' was added. This case currently does nothing (break;) and can be extended in the future. The reason for this addition is to prepare the WebSocketDebugger to handle routing cases as we evolved the system to support new WebSocket features.