"Added new lines to achieve better code readability
In this commit, newline characters were added into several PHP files in the CloudAdmin project for better code readability. This was done to separate the license header from the rest of the code for better readability by developers. No change in actual functionality occured. The change affects almost all of the files in the project, making it more uniform and easier to read across all files."