Introduce explicit global function invocation
The commit introduces explicit global function invocation to the codebase. PHP’s performance could be potentially improved by these changes because when a function is called, it will check the namespace first, if the namespace is not defined it will fallback to check the global function namespace. By doing this explicit call we save this check.

Function calls have been edited to explicitly call global functions across various files.

Additionally, this commit introduces the 'native_function_invocation' option in the .php-cs-fixer.php file. This will enforce the use of explicit calls to global functions anywhere in the codebase.

Lastly, refactoring codebase to integrate explicit calls to allow PHP to resolve global functions faster.