Add FlameGraphChart to MenuController
Added a new path for the 'flameGraphChart' component in the MenuController. This change will allow users to access the 'flameGraphChart' view directly from the menu. This makes the application more user-friendly, as they no longer need to search for this particular chart elsewhere. The relevant meta properties like 'icon', 'title', 'isLink', 'isHide', 'isFull', 'isAffix', 'isKeepAlive' are also defined to ensure proper behavior and depiction in the UI.