Add newline at the end of todo list
The lack of a newline character at the end of the todo list file (ie- xhgui+xhprof(进行中)\ No newline at end of file is changed to xhgui+xhprof) was causing the file to be incorrectly parsed. This commit adds that newline to ensure consistent parsing across various editors and systems.