Refactor ProfilerMiddleware and update ignore files
In ProfilerMiddleware, the usage of several Profiler event handlers (CalculateDiffsBetweenEdges, CleanupEvent, PrepareEdges, and PreparePeaks) were removed, and refactored the starting of profiling to occur only when it's enabled. This makes the profiling process more efficient by starting profiling only when necessary.

In .php-cs-fixer.php and .gitignore, 'hyperf-casbin-example' was added to the list of excluded directories. This is necessary to prevent unnecessary attention to test directories that doesn't contribute to the main application.