Add new RoleController and update MenuController
In this commit, a new RoleController is added to the Sys directory, that encompasses list, add and delete functions. This controller is essential for the management of user roles within the application, hence serving as an important feature for user access control.

In addition, the MenuController within the same directory is also updated. Annotation for Middleware is moved to class level to ensure authentication is required for all routes. Add, update, and delete post routes were added to enhance functionality for our menu management.