Refactor models to use CloudAdmin base model
Deleted the existing base model as it was redundant and switched all usage from App\Model\Model to CloudAdmin\Model\Model. This change was necessary to streamline the code with the use of a shared base model from CloudAdmin for better code maintenance and reuse. Additional constants were introduced in SystemMenu model to address magic values. The 'getRules' function in FormRequest now falls back to an empty array when no scene is provided.