Implement 'add' function in MenuController and refactor MenuService
In the MenuController, the 'add' function was implemented for accepting 'MenuRequest' and processing it. This function now uses 'MenuLogic' to process the request and generate a success response. An instance of MenuLogic was injected using the Hyperf 'Inject' annotation.

In MenuService, the 'add' method was re-implemented to create and return a SystemMenu model based on the input data. The old implementation of the add function and the generateCrudButton function was removed.

The Menu logic class has been refactored. The 'add' method was implemented here