Update task statuses in todo.list
The todo.list file was updated to reflect the current status of various tasks. The task '*use Box to package this project*' is now marked as pending development. The tasks '*xhgui+xhprof*' and '*page pagination feature*' have updated statuses, with the former set as completed and the latter as in progress. This was done to ensure clarity on the project's progression.