Update debug handler and add authorization check to WebSocketDebugger
The 'handler' class was updated from Swow\Debug\Debugger\Debugger to CloudAdmin\SDB\WebSocketDebugger to utilize the latter's capabilities. Also added an authorization check to WebSocketDebugger to enhance security. The system now checks the 'authorization' header of each incoming HTTP request, decodes the base64-encoded username and password, and returns an unauthorized error if the