Add basic auth to WebSocketDebugger
In this commit, Basic Username and Basic Password were added to the .env file, and these new environment variables are used for authorization checks in the WebSocketDebugger class. This change was made to allow for a secure way to check authorizations during WebSocket connections. Toggle support was also added by checking for an 'ENABLE_BASIC' environment variable to choose to use Basic Auth or not.