Add tree menu functionality and update dependencies
Added 'aimeos/map' package to composer.json for better map handling.
Introduced a new 'tree' function in the MenuController, MenuService, and MenuLogic for displaying menu in tree structure.
Updated the handler for debugger configuration to use Swow Debugger class.
Commented out the ProfilerMiddleware in the middlewares configuration to disable profiling.
This is geared towards improving navigation and overall user experience.