Enable authentication middleware and update menu listing logic
Middleware for authentication has been uncommented in MenuController to ensure secured access. Also, the logic for listing menus in MenuLogic is updated so not only parent menus but also other menus based on provided parameters can be obtained with pagination. These changes are made for enhanced security and flexibility in menu listing.