Add a variable for error message source in BusinessException
A new variable errMsgIsFromErrorCode is added to BusinessException class to flag if error messages are from error code. If true, the handler responds with failure status and the error message, rather than just the error code. This change was made to provide more detailed error messages to the end user. In addition, in the System/MenuLogic.php file, a BusinessException is thrown for exceptions related to failed update operations.