Rename 'is_hidden' to 'is_display' and update method calls
The attribute `is_hidden` in the `SystemMenu.php` model was renamed to `is_display` for clearer understanding. This attribute's definition and references were updated in all related files such as `MenuController.php` and `MenuLogic.php`. Also, the `list` method in these files was refactored to include a parameter variable for increased readability and flexibility. Furthermore, a new private method `filterParams` was introduced in `MenuLogic.php` to handle parameter filtering.